feat(06-01): decompose edit page monolith into section components with shared hook
- Extract useCommunityPageForm hook with all state, queries, mutations, and actions
- Create 5 section components: MainSettings, Bio, Gallery, Services, Pricing
- Rewrite community-edit-content.tsx as 143-line thin orchestrator (was 1415)
- Sticky header bar with Save/Publish buttons always visible
- Zero visual regression: identical render output, different file organization
